pyproject.toml
[project]
name = "ansible-jenkins"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-jenkins>=1.8.3",
]
